py_translate/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
flask==3.0.3
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
python-dotenv==1.1.0
    # via -r requirements.in
waitress==3.0.0
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
